add NISR reweighting for signal samples
The SMS template maker now computes the total NISR weight and rescales
the sample to keep the cross section constant.  Note: does this
regardless of whether or not the NISR weighting is actually applied --
should modify at some point to only rescale the sample if the weights
are turned on.