Fix limit file parsing in corner case
There was an issue parsing ntuple and limit files in the case that
one SMS name was a substring of a different SMS name.  Fixed this
and also removed the T5ttcc extension as a submodel because the hadd
code already renames the extension ntuple files correctly.
Also added T5qqqqVV to the list of models.